clearing up the mess 	-instruction sets are redesigned 	-new 64kb memory and 16bit addressing system added

+++ src/Memory.cpp
@@ -1,8 +1,8 @@
 #include "../include/Memory.h"
-void Memory::write(uint8_t address,uint8_t data){
+void Memory::write(uint16_t address,uint8_t data){
      this->data[address]=data;
 }
-uint8_t Memory::read(uint8_t address) const{
+uint8_t Memory::read(uint16_t address) const{
      return data[address];
  }
 void Memory::Display() const{
